pr fixies, rename auth service file
