+++ Dockerfile
@@ -1,11 +1,11 @@
-FROM node:lts-alpine as build-stage
+FROM node:14-alpine as build-stage
 WORKDIR /app
 COPY package*.json ./
 RUN HUSKY_SKIP_INSTALL=true npm install --unsafe-perm
 COPY . .
 RUN npm run build-prod
 
-FROM nginx:stable-alpine as production-stage
+FROM nginx:1.19.4-alpine as production-stage
 COPY --from=build-stage /app/dist /usr/share/nginx/html
 COPY --from=build-stage /app/nginx.conf.template /etc/nginx/conf.d
 RUN rm /etc/nginx/conf.d/default.conf
