+++ client/services/auth/use-auth-service.ts
